fix: Pass model_architecture to import_torch_model if model is pytorch type
